Remove remaining traces of changelog
We intentionally stopped generating a changelog from git history in the
switch to meson, since the tiny number of people who care can just look
at the git history directly. For end users the "NEWS" file is a more
consumable record of what's changed at a high level.

The empty ChangeLog was still in git though from the autoconf days and
thus ended up in the meson generated dist. It was also still mentioned
in syntax check rules and the RPM spec.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>