src: Fix ssh tunnel with non-POSIX remote shells
The command depends on POSIX-compatible shell being the default shell on
the remote side of SSH, but that might not be the case.  To make sure
the command gets parsed correctly this commit encloses it in extra
single quotes (to avoid it being parsed by the remote shell) and passes
that string as a parameter to `sh -c`.

Signed-off-by: Martin Kletzander <nert.pinx@gmail.com>