Make app-id and desktop file name match
The application uses an application-id of
"org.virt-manager.virt-viewer" (see function
remote_viewer_new in src/remote-viewer.c).

So far, there was a mismatch between this
application-id and the desktop file name,
resulting e.g. in no proper window icon being
used when running the app on KDE Plasma Wayland.

Adjust the name of the desktop and appdata
files to match the application-id as expected.

Also update the translatable strings using this
command:

    ninja -C _build virt-viewer-pot

Fixes: https://gitlab.com/virt-viewer/virt-viewer/-/issues/135
Signed-off-by: Michael Weghorn <m.weghorn@posteo.de>